finalised support for bitbucket piplines, make sure upgrade operation updates build scripts, dont create .pdqtest_old files any more, release

+++ lib/pdqtest/upgrade.rb
@@ -1,3 +1,5 @@
+require 'pdqtest/skeleton'
+
 module PDQTest
   module Upgrade
     GEMFILE             = 'Gemfile'
@@ -13,6 +15,10 @@ module PDQTest
         'line'  => "gem 'puppet-strings', :git => 'https://github.com/puppetlabs/puppet-strings'",
         'added' => false,
       },
+      'puppet'          => {
+        'line'  => "gem 'puppet', '#{PDQTest::PUPPET_VERSION}'",
+        'added' => false,
+      }
     }
 
 
@@ -20,43 +26,51 @@ module PDQTest
     def self.upgrade()
       t_file = File.open("#{GEMFILE}.tmp","w")
       updating_gem = false
-      File.open(GEMFILE, 'r') do |f|
-        f.each_line{ |line|
-          if line =~ GEM_REGEXP
-            # a gem stanza
-            processing_gem = $2
-            if GEMS.keys.include?(processing_gem)
-              # fixup one of our monitored gems as needed, mark
-              # this as being a gem that is being updated so
-              # that we can kill any multi-line attributes
-              t_file.puts GEMS[processing_gem]['line']
-              updating_gem = true
-              GEMS[processing_gem]['added'] = true
+
+      if File.exists?(GEMFILE)
+        File.open(GEMFILE, 'r') do |f|
+          f.each_line{ |line|
+            if line =~ GEM_REGEXP
+              # a gem stanza
+              processing_gem = $2
+              if GEMS.keys.include?(processing_gem)
+                # fixup one of our monitored gems as needed, mark
+                # this as being a gem that is being updated so
+                # that we can kill any multi-line attributes
+                t_file.puts GEMS[processing_gem]['line']
+                updating_gem = true
+                GEMS[processing_gem]['added'] = true
+              else
+                # a gem we don't care about - write it out as-is
+                t_file.puts line
+                updating_gem = false
+              end
+            elsif updating_gem and line =~ GEM_ATTRIB_REGEXP
+              # do nothing - remove the multi-line attributes
             else
-              # a gem we don't care about - write it out as-is
+              # anything else... (esp comments)
               t_file.puts line
-              updating_gem = false
             end
-          elsif updating_gem and line =~ GEM_ATTRIB_REGEXP
-            # do nothing - remove the multi-line attributes
-          else
-            # anything else... (esp comments)
-            t_file.puts line
+          }
+        end
+
+        # the code above will only UPGRADE existing gem lines, but if this is our
+        # first run, there will be nothing to upgrade, so loop through the GEMS
+        # for any that are not already added and append them
+        GEMS.each { |name, opts|
+          if ! opts['added']
+            t_file.puts opts['line']
           end
         }
+
+        t_file.close
+        FileUtils.mv(t_file.path, GEMFILE)
+
+        # miscellanious file updates
+        PDQTest::Skeleton::install_integrations()
+      else
+        Escort::Logger.error.error "Gemfile not found in #{Dir.pwd}"
       end
-
-      # the code above will only UPGRADE existing gem lines, but if this is our
-      # first run, there will be nothing to upgrade, so loop through the GEMS
-      # for any that are not already added and append them
-      GEMS.each { |name, opts|
-        if ! opts['added']
-          t_file.puts opts['line']
-        end
-      }
-
-      t_file.close
-      FileUtils.mv(t_file.path, GEMFILE)
     end
 
   end
